chore(notebook): delete notebook_state.rs and dead receiver tasks
Delete crates/notebook/src/notebook_state.rs entirely (-753 lines).
The NotebookState struct, FrontendCell enum, and all constructors
(new_empty_with_runtime, from_notebook, etc.) are no longer used
by any code in the workspace.

The only live dependency was snapshot_from_nbformat(), used by
import_pyproject_dependencies and import_pixi_dependencies. Moved
to lib.rs as a local function.

Also removed:
- Dead SyncUpdate receiver tasks from setup_sync_receivers and
  initialize_notebook_sync. In pipe mode the relay never sends
  metadata diffs through changes_tx, and no frontend code listens
  for the notebook:metadata_updated event.
- Unused NotebookSyncReceiver import.

Net: -838 lines.
