--- crates/notebook/src/notebook_state.rs
@@ -1,753 +1,0 @@
-use crate::runtime::Runtime;
-use crate::settings::{self, PythonEnvType};
-use log::info;
-use nbformat::v4::{Cell, CellId, CellMetadata, Notebook};
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
-use std::path::PathBuf;
-use uuid::Uuid;
-
-/// Migrate legacy metadata format to the new `runt` namespace structure.
-///
-/// Old format:
-///   metadata.uv.dependencies, metadata.conda.dependencies, metadata.deno
-///
-/// New format:
-///   metadata.runt.uv.dependencies, metadata.runt.conda.dependencies, metadata.runt.deno
-///
-/// Returns `true` if any migration was performed.
-pub fn migrate_legacy_metadata(additional: &mut HashMap<String, serde_json::Value>) -> bool {
-    let mut migrated = false;
-
-    // Extract legacy keys first (before borrowing runt mutably)
-    let legacy_uv = additional.remove("uv");
-    let legacy_conda = additional.remove("conda");
-    let legacy_deno = additional.remove("deno");
-
-    if legacy_uv.is_none() && legacy_conda.is_none() && legacy_deno.is_none() {
-        return false;
-    }
-
-    // Get or create runt namespace
-    let runt = additional
-        .entry("runt".to_string())
-        .or_insert_with(|| serde_json::json!({"schema_version": "1"}));
-
-    let runt_obj = match runt.as_object_mut() {
-        Some(obj) => obj,
-        None => {
-            // Put the keys back if we can't migrate
-            if let Some(uv) = legacy_uv {
-                additional.insert("uv".to_string(), uv);
-            }
-            if let Some(conda) = legacy_conda {
-                additional.insert("conda".to_string(), conda);
-            }
-            if let Some(deno) = legacy_deno {
-                additional.insert("deno".to_string(), deno);
-            }
-            return false;
-        }
-    };
-
-    // Migrate uv if present and not already in runt
-    if let Some(uv) = legacy_uv {
-        if !runt_obj.contains_key("uv") {
-            runt_obj.insert("uv".to_string(), uv);
-            migrated = true;
-            info!("[metadata-migration] Migrated metadata.uv -> metadata.runt.uv");
-        }
-    }
-
-    // Migrate conda if present and not already in runt
-    if let Some(conda) = legacy_conda {
-        if !runt_obj.contains_key("conda") {
-            runt_obj.insert("conda".to_string(), conda);
-            migrated = true;
-            info!("[metadata-migration] Migrated metadata.conda -> metadata.runt.conda");
-        }
-    }
-
-    // Migrate deno if present and not already in runt
-    if let Some(deno) = legacy_deno {
-        if !runt_obj.contains_key("deno") {
-            runt_obj.insert("deno".to_string(), deno);
-            migrated = true;
-            info!("[metadata-migration] Migrated metadata.deno -> metadata.runt.deno");
-        }
-    }
-
-    migrated
-}
-
-/// Flattened cell representation for the frontend.
-/// Converts nbformat's tagged enum into something JS-friendly.
-#[derive(Serialize, Deserialize, Clone, Debug)]
-#[serde(tag = "cell_type")]
-pub enum FrontendCell {
-    #[serde(rename = "code")]
-    Code {
-        id: String,
-        source: String,
-        execution_count: Option<i32>,
-        outputs: Vec<serde_json::Value>,
-    },
-    #[serde(rename = "markdown")]
-    Markdown { id: String, source: String },
-    #[serde(rename = "raw")]
-    Raw { id: String, source: String },
-}
-
-impl FrontendCell {
-    pub fn id(&self) -> &str {
-        match self {
-            FrontendCell::Code { id, .. } => id,
-            FrontendCell::Markdown { id, .. } => id,
-            FrontendCell::Raw { id, .. } => id,
-        }
-    }
-}
-
-/// Convert nbformat Cell -> FrontendCell
-fn cell_to_frontend(cell: &Cell) -> FrontendCell {
-    match cell {
-        Cell::Code {
-            id,
-            source,
-            execution_count,
-            outputs,
-            ..
-        } => FrontendCell::Code {
-            id: id.to_string(),
-            source: source.join(""),
-            execution_count: *execution_count,
-            outputs: outputs
-                .iter()
-                .filter_map(|o| serde_json::to_value(o).ok())
-                .collect(),
-        },
-        Cell::Markdown { id, source, .. } => FrontendCell::Markdown {
-            id: id.to_string(),
-            source: source.join(""),
-        },
-        Cell::Raw { id, source, .. } => FrontendCell::Raw {
-            id: id.to_string(),
-            source: source.join(""),
-        },
-    }
-}
-
-pub struct NotebookState {
-    pub notebook: Notebook,
-    pub path: Option<PathBuf>,
-    pub dirty: bool,
-    /// Working directory for untitled notebooks (used for project file detection).
-    /// Set from --cwd CLI argument and preserved across daemon reconnects.
-    pub working_dir: Option<PathBuf>,
-}
-
-impl NotebookState {
-    /// Create a new empty notebook with a specific runtime
-    pub fn new_empty_with_runtime(runtime: Runtime) -> Self {
-        let env_id = Uuid::new_v4().to_string();
-        let mut additional = HashMap::new();
-
-        // Build runt metadata with nested env config based on runtime
-        let runt_meta = match &runtime {
-            Runtime::Python => {
-                // Load user's preferred Python environment type from settings
-                let app_settings = settings::load_settings();
-                match app_settings.default_python_env {
-                    PythonEnvType::Uv | PythonEnvType::Other(_) => {
-                        serde_json::json!({
-                            "schema_version": "1",
-                            "env_id": env_id,
-                            "uv": {
-                                "dependencies": Vec::<String>::new(),
-                            }
-                        })
-                    }
-                    PythonEnvType::Conda => {
-                        serde_json::json!({
-                            "schema_version": "1",
-                            "env_id": env_id,
-                            "conda": {
-                                "dependencies": Vec::<String>::new(),
-                                "channels": vec!["conda-forge"],
-                            }
-                        })
-                    }
-                }
-            }
-            Runtime::Deno => {
-                // Deno setup with default permissions
-                // Note: deno permissions stay at top level since they're not our namespace
-                additional.insert(
-                    "deno".to_string(),
-                    serde_json::json!({
-                        "permissions": Vec::<String>::new(),
-                    }),
-                );
-                serde_json::json!({
-                    "schema_version": "1",
-                    "env_id": env_id,
-                })
-            }
-            Runtime::Other(_) => {
-                // Unknown runtime — just store schema version and env_id
-                serde_json::json!({
-                    "schema_version": "1",
-                    "env_id": env_id,
-                })
-            }
-        };
-
-        additional.insert("runt".to_string(), runt_meta);
-
-        // Set standard Jupyter kernelspec based on runtime
-        // This is critical for daemon detection of kernel type
-        let kernelspec = match &runtime {
-            Runtime::Python => Some(nbformat::v4::KernelSpec {
-                name: "python3".to_string(),
-                display_name: "Python 3".to_string(),
-                language: Some("python".to_string()),
-                additional: std::collections::HashMap::new(),
-            }),
-            Runtime::Deno => Some(nbformat::v4::KernelSpec {
-                name: "deno".to_string(),
-                display_name: "Deno".to_string(),
-                language: Some("typescript".to_string()),
-                additional: std::collections::HashMap::new(),
-            }),
-            Runtime::Other(s) => Some(nbformat::v4::KernelSpec {
-                name: s.clone(),
-                display_name: s.clone(),
-                language: None,
-                additional: std::collections::HashMap::new(),
-            }),
-        };
-
-        NotebookState {
-            notebook: Notebook {
-                metadata: nbformat::v4::Metadata {
-                    kernelspec,
-                    language_info: None,
-                    authors: None,
-                    additional,
-                },
-                nbformat: 4,
-                nbformat_minor: 5,
-                cells: vec![Cell::Code {
-                    id: CellId::from(Uuid::new_v4()),
-                    metadata: empty_cell_metadata(),
-                    execution_count: None,
-                    source: Vec::new(),
-                    outputs: Vec::new(),
-                }],
-            },
-            path: None,
-            dirty: false,
-            working_dir: None,
-        }
-    }
-
-    /// Create a new empty Python notebook with conda metadata from an environment.yml config.
-    /// Used when creating a new notebook in a directory with environment.yml.
-    /// Unlike pyproject.toml, deps are NOT copied into the notebook — they live in the project file.
-    /// We just set conda mode with the environment.yml's channels so the UI shows correctly.
-    pub fn new_empty_with_conda_from_environment_yml(
-        config: &crate::environment_yml::EnvironmentYmlConfig,
-    ) -> Self {
-        let env_id = Uuid::new_v4().to_string();
-        let mut additional = HashMap::new();
-
-        let channels = if config.channels.is_empty() {
-            vec!["conda-forge".to_string()]
-        } else {
-            config.channels.clone()
-        };
-
-        additional.insert(
-            "runt".to_string(),
-            serde_json::json!({
-                "schema_version": "1",
-                "env_id": env_id,
-                "conda": {
-                    "dependencies": Vec::<String>::new(),
-                    "channels": channels,
-                }
-            }),
-        );
-
-        // Python kernelspec for conda from environment.yml
-        let kernelspec = Some(nbformat::v4::KernelSpec {
-            name: "python3".to_string(),
-            display_name: "Python 3".to_string(),
-            language: Some("python".to_string()),
-            additional: std::collections::HashMap::new(),
-        });
-
-        NotebookState {
-            notebook: Notebook {
-                metadata: nbformat::v4::Metadata {
-                    kernelspec,
-                    language_info: None,
-                    authors: None,
-                    additional,
-                },
-                nbformat: 4,
-                nbformat_minor: 5,
-                cells: vec![Cell::Code {
-                    id: CellId::from(Uuid::new_v4()),
-                    metadata: empty_cell_metadata(),
-                    execution_count: None,
-                    source: Vec::new(),
-                    outputs: Vec::new(),
-                }],
-            },
-            path: None,
-            dirty: false,
-            working_dir: None,
-        }
-    }
-
-    /// Create a new empty Python notebook with UV metadata from a pyproject.toml config.
-    /// Used when creating a new notebook in a directory with pyproject.toml.
-    pub fn new_empty_with_uv_from_pyproject(config: &crate::pyproject::PyProjectConfig) -> Self {
-        let env_id = Uuid::new_v4().to_string();
-        let mut additional = HashMap::new();
-
-        let all_deps = crate::pyproject::get_all_dependencies(config);
-
-        additional.insert(
-            "runt".to_string(),
-            serde_json::json!({
-                "schema_version": "1",
-                "env_id": env_id,
-                "uv": {
-                    "dependencies": all_deps,
-                    "requires-python": config.requires_python,
-                }
-            }),
-        );
-
-        // Python kernelspec for UV from pyproject.toml
-        let kernelspec = Some(nbformat::v4::KernelSpec {
-            name: "python3".to_string(),
-            display_name: "Python 3".to_string(),
-            language: Some("python".to_string()),
-            additional: std::collections::HashMap::new(),
-        });
-
-        NotebookState {
-            notebook: Notebook {
-                metadata: nbformat::v4::Metadata {
-                    kernelspec,
-                    language_info: None,
-                    authors: None,
-                    additional,
-                },
-                nbformat: 4,
-                nbformat_minor: 5,
-                cells: vec![Cell::Code {
-                    id: CellId::from(Uuid::new_v4()),
-                    metadata: empty_cell_metadata(),
-                    execution_count: None,
-                    source: Vec::new(),
-                    outputs: Vec::new(),
-                }],
-            },
-            path: None,
-            dirty: false,
-            working_dir: None,
-        }
-    }
-
-    pub fn from_notebook(notebook: Notebook, path: PathBuf) -> Self {
-        NotebookState {
-            notebook,
-            path: Some(path),
-            dirty: false,
-            working_dir: None, // Saved notebooks use path for project detection
-        }
-    }
-
-    /// Get the runtime type from notebook metadata.
-    ///
-    /// Reads from kernelspec.name (the standard Jupyter field), not runt.runtime.
-    /// This mirrors the daemon's `detect_notebook_kernel_type()` logic so
-    /// notebook-side and daemon-side detection stay consistent.
-    pub fn get_runtime(&self) -> Runtime {
-        // Check kernelspec.name first (most reliable)
-        if let Some(ks) = &self.notebook.metadata.kernelspec {
-            let name_lower = ks.name.to_lowercase();
-            if name_lower.contains("deno") {
-                return Runtime::Deno;
-            }
-            if name_lower.contains("python") {
-                return Runtime::Python;
-            }
-            // Check language field
-            if let Some(lang) = &ks.language {
-                let lang_lower = lang.to_lowercase();
-                if lang_lower == "typescript" || lang_lower == "javascript" {
-                    return Runtime::Deno;
-                }
-                if lang_lower == "python" {
-                    return Runtime::Python;
-                }
-            }
-            // Unknown kernelspec - preserve as Other for round-tripping
-            return Runtime::Other(ks.name.clone());
-        }
-
-        // Fallback: check language_info.name
-        if let Some(lang_info) = &self.notebook.metadata.language_info {
-            let name_lower = lang_info.name.to_lowercase();
-            if name_lower == "typescript" || name_lower == "javascript" || name_lower == "deno" {
-                return Runtime::Deno;
-            }
-            if name_lower == "python" {
-                return Runtime::Python;
-            }
-        }
-
-        // Default to Python only if no kernelspec and no language_info
-        Runtime::Python
-    }
-
-    pub fn cells_for_frontend(&self) -> Vec<FrontendCell> {
-        self.notebook.cells.iter().map(cell_to_frontend).collect()
-    }
-}
-
-// ── Conversions between nbformat Metadata and NotebookMetadataSnapshot ──
-
-use runtimed::notebook_metadata::{
-    CondaInlineMetadata, DenoMetadata, KernelspecSnapshot, LanguageInfoSnapshot,
-    NotebookMetadataSnapshot, RuntMetadata, UvInlineMetadata,
-};
-
-/// Extract a `NotebookMetadataSnapshot` from nbformat metadata.
-///
-/// Reads `kernelspec`, `language_info`, and `runt` from the nbformat struct.
-/// Falls back to legacy `uv`/`conda` top-level keys if `runt` is absent.
-pub fn snapshot_from_nbformat(metadata: &nbformat::v4::Metadata) -> NotebookMetadataSnapshot {
-    let kernelspec = metadata.kernelspec.as_ref().map(|ks| KernelspecSnapshot {
-        name: ks.name.clone(),
-        display_name: ks.display_name.clone(),
-        language: ks.language.clone(),
-    });
-
-    let language_info = metadata
-        .language_info
-        .as_ref()
-        .map(|li| LanguageInfoSnapshot {
-            name: li.name.clone(),
-            version: li.version.clone(),
-        });
-
-    // Helper to convert legacy DenoDependencies to DenoMetadata
-    let convert_legacy_deno = |dd: crate::deno_env::DenoDependencies| DenoMetadata {
-        permissions: dd.permissions,
-        import_map: dd.import_map,
-        config: dd.config,
-        // Only record if non-default (default is true)
-        flexible_npm_imports: if dd.flexible_npm_imports {
-            None
-        } else {
-            Some(false)
-        },
-    };
-
-    let runt = if let Some(runt_value) = metadata.additional.get("runt") {
-        let mut runt_meta = serde_json::from_value::<RuntMetadata>(runt_value.clone())
-            .unwrap_or_else(|_| RuntMetadata {
-                schema_version: "1".to_string(),
-                env_id: None,
-                uv: None,
-                conda: None,
-                deno: None,
-                trust_signature: None,
-                trust_timestamp: None,
-            });
-
-        // Also check legacy top-level "deno" key - this is where the Tauri commands write
-        // deno metadata, so we need to merge it even when runt exists
-        if runt_meta.deno.is_none() {
-            if let Some(deno_value) = metadata.additional.get("deno") {
-                if let Ok(dd) =
-                    serde_json::from_value::<crate::deno_env::DenoDependencies>(deno_value.clone())
-                {
-                    runt_meta.deno = Some(convert_legacy_deno(dd));
-                }
-            }
-        }
-
-        runt_meta
-    } else {
-        // Fallback: legacy top-level uv/conda/deno keys
-        let uv = metadata
-            .additional
-            .get("uv")
-            .and_then(|v| serde_json::from_value::<UvInlineMetadata>(v.clone()).ok());
-        let conda = metadata
-            .additional
-            .get("conda")
-            .and_then(|v| serde_json::from_value::<CondaInlineMetadata>(v.clone()).ok());
-        let deno = metadata
-            .additional
-            .get("deno")
-            .and_then(|v| {
-                serde_json::from_value::<crate::deno_env::DenoDependencies>(v.clone()).ok()
-            })
-            .map(convert_legacy_deno);
-
-        RuntMetadata {
-            schema_version: "1".to_string(),
-            env_id: None,
-            uv,
-            conda,
-            deno,
-            trust_signature: None,
-            trust_timestamp: None,
-        }
-    };
-
-    NotebookMetadataSnapshot {
-        kernelspec,
-        language_info,
-        runt,
-    }
-}
-
-/// Merge a `NotebookMetadataSnapshot` back into nbformat metadata.
-///
-/// Replaces `kernelspec`, `language_info`, and `metadata.additional["runt"]`
-/// while preserving all other keys in `additional` (e.g. `jupyter`, `deno`
-/// top-level config, custom extensions).
-pub fn merge_snapshot_into_nbformat(
-    snapshot: &NotebookMetadataSnapshot,
-    metadata: &mut nbformat::v4::Metadata,
-) {
-    // Replace kernelspec, preserving existing additional fields
-    let existing_ks_additional = metadata
-        .kernelspec
-        .as_ref()
-        .map(|ks| ks.additional.clone())
-        .unwrap_or_default();
-    metadata.kernelspec = snapshot
-        .kernelspec
-        .as_ref()
-        .map(|ks| nbformat::v4::KernelSpec {
-            name: ks.name.clone(),
-            display_name: ks.display_name.clone(),
-            language: ks.language.clone(),
-            additional: existing_ks_additional,
-        });
-
-    // Replace language_info, preserving existing codemirror_mode and additional
-    let (existing_cm_mode, existing_li_additional) = metadata
-        .language_info
-        .as_ref()
-        .map(|li| (li.codemirror_mode.clone(), li.additional.clone()))
-        .unwrap_or_default();
-    metadata.language_info = snapshot
-        .language_info
-        .as_ref()
-        .map(|li| nbformat::v4::LanguageInfo {
-            name: li.name.clone(),
-            version: li.version.clone(),
-            codemirror_mode: existing_cm_mode,
-            additional: existing_li_additional,
-        });
-
-    // Deep-merge runt namespace to preserve unknown fields (e.g. trust_signature)
-    if let Ok(mut new_runt) = serde_json::to_value(&snapshot.runt) {
-        if let Some(existing_runt) = metadata.additional.get("runt") {
-            // Merge: start with existing, overwrite with new snapshot fields
-            if let (Some(existing_obj), Some(new_obj)) =
-                (existing_runt.as_object(), new_runt.as_object_mut())
-            {
-                for (k, v) in existing_obj {
-                    // Only keep existing keys that aren't in the new snapshot
-                    if !new_obj.contains_key(k) {
-                        new_obj.insert(k.clone(), v.clone());
-                    }
-                }
-            }
-        }
-        metadata.additional.insert("runt".to_string(), new_runt);
-    }
-}
-
-fn empty_cell_metadata() -> CellMetadata {
-    CellMetadata {
-        id: None,
-        collapsed: None,
-        scrolled: None,
-        deletable: None,
-        editable: None,
-        format: None,
-        name: None,
-        tags: None,
-        jupyter: None,
-        execution: None,
-        additional: HashMap::new(),
-    }
-}
-
-#[cfg(test)]
-#[allow(clippy::unwrap_used, clippy::expect_used)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_new_empty_with_runtime_python() {
-        let state = NotebookState::new_empty_with_runtime(Runtime::Python);
-
-        // Should have runt namespace with either uv or conda nested inside
-        let runt = state
-            .notebook
-            .metadata
-            .additional
-            .get("runt")
-            .expect("runt namespace should exist");
-        let has_env = runt.get("uv").is_some() || runt.get("conda").is_some();
-        assert!(has_env);
-
-        // Runtime is determined by kernelspec, not runt.runtime
-        let ks = state.notebook.metadata.kernelspec.as_ref().unwrap();
-        assert!(ks.name.contains("python"));
-    }
-
-    #[test]
-    fn test_new_empty_with_runtime_deno() {
-        let state = NotebookState::new_empty_with_runtime(Runtime::Deno);
-
-        assert!(state.notebook.metadata.additional.contains_key("deno"));
-
-        // Runtime is determined by kernelspec, not runt.runtime
-        let ks = state.notebook.metadata.kernelspec.as_ref().unwrap();
-        assert_eq!(ks.name, "deno");
-    }
-
-    #[test]
-    fn test_get_runtime_returns_python_by_default() {
-        let state = NotebookState::new_empty_with_runtime(Runtime::Python);
-        assert_eq!(state.get_runtime(), Runtime::Python);
-    }
-
-    #[test]
-    fn test_get_runtime_returns_correct_runtime() {
-        let state = NotebookState::new_empty_with_runtime(Runtime::Deno);
-        assert_eq!(state.get_runtime(), Runtime::Deno);
-    }
-
-    #[test]
-    fn test_get_runtime_other_round_trips() {
-        // Unknown runtimes should round-trip through kernelspec
-        let state = NotebookState::new_empty_with_runtime(Runtime::Other("julia".into()));
-        assert_eq!(state.get_runtime(), Runtime::Other("julia".into()));
-    }
-
-    #[test]
-    fn test_get_runtime_contains_deno() {
-        // Should detect deno-variants like "deno-ts" via contains()
-        let mut state = NotebookState::new_empty_with_runtime(Runtime::Python);
-        state.notebook.metadata.kernelspec = Some(nbformat::v4::KernelSpec {
-            name: "deno-typescript".to_string(),
-            display_name: "Deno TypeScript".to_string(),
-            language: None,
-            additional: std::collections::HashMap::new(),
-        });
-        assert_eq!(state.get_runtime(), Runtime::Deno);
-    }
-
-    #[test]
-    fn test_get_runtime_language_info_fallback() {
-        // Should fall back to language_info.name when kernelspec is missing
-        let mut state = NotebookState::new_empty_with_runtime(Runtime::Python);
-        state.notebook.metadata.kernelspec = None;
-        state.notebook.metadata.language_info = Some(nbformat::v4::LanguageInfo {
-            name: "typescript".to_string(),
-            version: None,
-            codemirror_mode: None,
-            additional: std::collections::HashMap::new(),
-        });
-        assert_eq!(state.get_runtime(), Runtime::Deno);
-    }
-
-    #[test]
-    fn test_get_runtime_kernelspec_language_fallback() {
-        // Should detect via kernelspec.language when name doesn't match
-        let mut state = NotebookState::new_empty_with_runtime(Runtime::Python);
-        state.notebook.metadata.kernelspec = Some(nbformat::v4::KernelSpec {
-            name: "custom-kernel".to_string(),
-            display_name: "Custom".to_string(),
-            language: Some("typescript".to_string()),
-            additional: std::collections::HashMap::new(),
-        });
-        assert_eq!(state.get_runtime(), Runtime::Deno);
-    }
-
-    #[test]
-    fn test_cells_for_frontend_converts_correctly() {
-        let mut state = NotebookState::new_empty_with_runtime(Runtime::Python);
-        // Set source directly on the nbformat cell
-        if let Cell::Code {
-            source: ref mut s, ..
-        } = state.notebook.cells[0]
-        {
-            *s = vec!["x = 1".to_string()];
-        }
-
-        let frontend_cells = state.cells_for_frontend();
-
-        assert_eq!(frontend_cells.len(), 1);
-        if let FrontendCell::Code { source, .. } = &frontend_cells[0] {
-            assert_eq!(source, "x = 1");
-        } else {
-            panic!("Expected code cell");
-        }
-    }
-
-    #[test]
-    fn test_frontend_cell_id_method() {
-        let code_cell = FrontendCell::Code {
-            id: "code-123".to_string(),
-            source: String::new(),
-            execution_count: None,
-            outputs: vec![],
-        };
-        let md_cell = FrontendCell::Markdown {
-            id: "md-456".to_string(),
-            source: String::new(),
-        };
-        let raw_cell = FrontendCell::Raw {
-            id: "raw-789".to_string(),
-            source: String::new(),
-        };
-
-        assert_eq!(code_cell.id(), "code-123");
-        assert_eq!(md_cell.id(), "md-456");
-        assert_eq!(raw_cell.id(), "raw-789");
-    }
-
-    #[test]
-    fn test_frontend_cell_serialization() {
-        let cell = FrontendCell::Code {
-            id: "test-id".to_string(),
-            source: "print('hi')".to_string(),
-            execution_count: Some(1),
-            outputs: vec![],
-        };
-
-        let json = serde_json::to_value(&cell).unwrap();
-
-        assert_eq!(json["cell_type"], "code");
-        assert_eq!(json["id"], "test-id");
-        assert_eq!(json["source"], "print('hi')");
-        assert_eq!(json["execution_count"], 1);
-    }
-}
